Fixed testing (missing import statement). Also renamed TowerDefenseAppSpec.groovy due to spelling error.
